Fix parquet metadata compliance

+++ Plank/Writing/ColumnChunkMetadata.cs
@@ -6,7 +6,7 @@ internal struct ColumnChunkMetadata
 {
     internal long DataPageOffset;
     internal long DictionaryPageOffset;
-    internal uint ValueCount;
+    internal long ValueCount;
     internal long TotalUncompressedSize;
     internal long TotalCompressedSize;
     internal EncodingKind DataEncoding;
